+++ backend/src/main/java/com/celuveat/auth/infra/oauth/naver/client/NaverMemberClient.java
@@ -1,9 +1,9 @@
-package com.celuveat.auth.infra.oauth.naver;
+package com.celuveat.auth.infra.oauth.naver.client;
 
 import com.celuveat.auth.domain.OauthMember;
 import com.celuveat.auth.domain.OauthServerType;
 import com.celuveat.auth.domain.client.OauthMemberClient;
-import com.celuveat.auth.infra.oauth.naver.client.NaverApiClient;
+import com.celuveat.auth.infra.oauth.naver.NaverOauthConfig;
 import com.celuveat.auth.infra.oauth.naver.dto.NaverMemberResponse;
 import com.celuveat.auth.infra.oauth.naver.dto.NaverToken;
 import lombok.RequiredArgsConstructor;
